### 2024-07-16 #6992 @editorialbot
[PRE REVIEW]: openBURST: Real-time air surveillance simulation and analysis for active and passive sensors
**Submitting author:** @z-mathews (<a href="http://orcid.org/0000-0002-6562-5584">Zenon Mathews</a>)
**Repository:** https://github.com/swiss-armed-forces/openburst
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
