### 2024-07-26 #6992 @hugoledoux
@danielskatz It's not really my area, but I see above that 2 already rejected (because of this I guess) and I don't see whose an expert in this area here...

However, be aware that I am going on holidays next week for 2w, so handling this might be slower than normal.

If you're okay with this, I'm happy to be the editor.
