### 2024-07-26 #6992 @danielskatz
@hugoledoux - thanks for offering. I think I'll go ahead and take this one myself, as I guess I'm as close to the domain as anyone...
