### 2024-07-26 #6992 @danielskatz
👋 @hasantahir & @thippesha-d - Would one or both of you be willing to review this JOSS submission?
