### 2024-07-26 #6992 @danielskatz
👋 @IgorBaratta - Would you be willing to review this JOSS submission?
