### 2024-07-26 #6992 @rvg296
> 👋 @rvg296 - Would you be willing to review this JOSS submission?

Yes, would be willing to
