### 2024-07-29 #6992 @hasantahir
> 👋 @hasantahir & @thippesha-d - Would one or both of you be willing to review this JOSS submission?

I am happy to do so.
