### 2024-07-29 #6992 @thippesha-d
> 👋 @hasantahir & @thippesha-d - Would one or both of you be willing to review this JOSS submission?

@danielskatz Yes, I will do, let me know how can I access the draft
