### 2024-07-29 #6992 @IgorBaratta
> 👋 @IgorBaratta - Would you be willing to review this JOSS submission?

Thank you for the invitation; unfortunately, I don't have the bandwidth to do it at the moment.
