### 2024-08-07 #7052 @danielskatz
👋 @rvg296 - you also should use the command `@editorialbot generate my checklist` to create your review checklist. @editorialbot commands need to be the first thing in a new comment.
