### 2024-09-02 Swiss-Armed-Forces/openburst#1 @hasantahir
JOSS Review - Compile and Test Issues
Hi,

I have currently reviewing your submission to JOSS and for that I wanted to test the tool. Whilst installing using the `pyproject.toml` file, I received a couple of errors:

```
pip._vendor.tomli._parser.TOMLDecodeError: Duplicate inline table key "name" (at line 8, column 91)
```
which I got through by removing the `author` description from the `[project]` section.

I then received the following error:

```shell
(openburst-venv) hasan@hasans:~/openburst$ python3.10 -m pip install .
Processing /home/hasan/openburst
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
  error: subprocess-exited-with-error

  × Getting requirements to build wheel did not run successfully.
  │ exit code: 1
  ╰─> [33 lines of output]
      Traceback (most recent call last):
        File "/home/hasan/openburst/openburst-venv/lib/python3.10/site-packages/pip/_vendor/pep517/in_process/_in_process.py", line 363, in <module>
          main()
        File "/home/hasan/openburst/openburst-venv/lib/python3.10/site-packages/pip/_vendor/pep517/in_process/_in_process.py", line 345, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
        File "/home/hasan/openburst/openburst-venv/lib/python3.10/site-packages/pip/_vendor/pep517/in_process/_in_process.py", line 130, in get_requires_for_build_wheel
          return hook(config_settings)
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/setuptools/build_meta.py", line 332, in get_requires_for_build_wheel
          return self._get_build_requires(config_settings, requirements=[])
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/setuptools/build_meta.py", line 302, in _get_build_requires
          self.run_setup()
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/setuptools/build_meta.py", line 318, in run_setup
          exec(code, locals())
        File "<string>", line 1, in <module>
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/setuptools/__init__.py", line 117, in setup
          return distutils.core.setup(**attrs)
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/setuptools/_distutils/core.py", line 158, in setup
          dist.parse_config_files()
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/setuptools/dist.py", line 608, in parse_config_files
          pyprojecttoml.apply_configuration(self, filename, ignore_option_errors)
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/setuptools/config/pyprojecttoml.py", line 72, in apply_configuration
          return _apply(dist, config, filepath)
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/setuptools/config/_apply_pyprojecttoml.py", line 50, in apply
          _apply_project_table(dist, config, root_dir)
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/setuptools/config/_apply_pyprojecttoml.py", line 76, in _apply_project_table
          corresp(dist, value, root_dir)
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/setuptools/config/_apply_pyprojecttoml.py", line 206, in _python_requires
          _set_config(dist, "python_requires", SpecifierSet(val))
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/packaging/specifiers.py", line 718, in __init__
          self._specs = frozenset(map(Specifier, split_specifiers))
        File "/tmp/pip-build-env-_kd91lvj/overlay/lib/python3.10/site-packages/packaging/specifiers.py", line 237, in __init__
          raise InvalidSpecifier(f"Invalid specifier: '{spec}'")
      packaging.specifiers.InvalidSpecifier: Invalid specifier: '3.10'
      [end of output]

error: subprocess-exited-with-error
```

which I circumvented by adding the following in the `pyproject.toml`:

```shell
[requires]
python_version = "3.10"
[project]
name = "openburst"
description = "A target and sensor simulator for air surveillance"
version = "1.0"
readme = "README.md"

```

I was able to install the tool but then get the following error whilst testing:

```shell
(openburst-venv) hasan@hasans:~/openburst$ pytest --pyargs openburst
========================================================= test session starts ==========================================================
rootdir: /home/hasan/openburst
configfile: pyproject.toml
collected 2 items / 5 errors

================================================================ ERRORS ================================================================
___________________ ERROR collecting openburst-venv/lib/python3.10/site-packages/openburst/tests/db/test_db_conn.py ____________________
ImportError while importing test module '/home/hasan/openburst/openburst-venv/lib/python3.10/site-packages/openburst/tests/db/test_db_conn.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst-venv/lib/python3.10/site-packages/openburst/tests/db/test_db_conn.py:6: in <module>
    from openburst.functions import dbfunctions
openburst-venv/lib/python3.10/site-packages/openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
_____________ ERROR collecting openburst-venv/lib/python3.10/site-packages/openburst/tests/dem/test_get_terrain_height.py ______________
ImportError while importing test module '/home/hasan/openburst/openburst-venv/lib/python3.10/site-packages/openburst/tests/dem/test_get_terrain_height.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst-venv/lib/python3.10/site-packages/openburst/tests/dem/test_get_terrain_height.py:8: in <module>
    from openburst.functions import basefunctions
openburst-venv/lib/python3.10/site-packages/openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst-venv/lib/python3.10/site-packages/openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
___________ ERROR collecting openburst-venv/lib/python3.10/site-packages/openburst/tests/openburst_splat/test_splat_load.py ____________
ImportError while importing test module '/home/hasan/openburst/openburst-venv/lib/python3.10/site-packages/openburst/tests/openburst_splat/test_splat_load.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst-venv/lib/python3.10/site-packages/openburst/tests/openburst_splat/test_splat_load.py:8: in <module>
    from openburst.functions import basefunctions
openburst-venv/lib/python3.10/site-packages/openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst-venv/lib/python3.10/site-packages/openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
______ ERROR collecting openburst-venv/lib/python3.10/site-packages/openburst/tests/openburst_splat/test_splat_parallel_matrix.py ______
ImportError while importing test module '/home/hasan/openburst/openburst-venv/lib/python3.10/site-packages/openburst/tests/openburst_splat/test_splat_parallel_matrix.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst-venv/lib/python3.10/site-packages/openburst/tests/openburst_splat/test_splat_parallel_matrix.py:11: in <module>
    from openburst.functions import basefunctions
openburst-venv/lib/python3.10/site-packages/openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst-venv/lib/python3.10/site-packages/openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
_____________ ERROR collecting openburst-venv/lib/python3.10/site-packages/openburst/tests/pcl/test_read_pcl_detections.py _____________
ImportError while importing test module '/home/hasan/openburst/openburst-venv/lib/python3.10/site-packages/openburst/tests/pcl/test_read_pcl_detections.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst-venv/lib/python3.10/site-packages/openburst/tests/pcl/test_read_pcl_detections.py:7: in <module>
    from openburst.functions import dbfunctions
openburst-venv/lib/python3.10/site-packages/openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
======================================================= short test summary info ========================================================
ERROR openburst-venv/lib/python3.10/site-packages/openburst/tests/db/test_db_conn.py
ERROR openburst-venv/lib/python3.10/site-packages/openburst/tests/dem/test_get_terrain_height.py
ERROR openburst-venv/lib/python3.10/site-packages/openburst/tests/openburst_splat/test_splat_load.py
ERROR openburst-venv/lib/python3.10/site-packages/openburst/tests/openburst_splat/test_splat_parallel_matrix.py
ERROR openburst-venv/lib/python3.10/site-packages/openburst/tests/pcl/test_read_pcl_detections.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 5 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
========================================================== 5 errors in 0.31s ===========================================================
```

Will appreciate feedback on this so that I can go ahead with the review.

Thanks
