### 2024-09-02 Swiss-Armed-Forces/openburst#1 @danielskatz
tagging for tracking (@hasantahir - please do this in future issues): https://github.com/openjournals/joss-reviews/issues/7052
