### 2024-09-03 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir  thanks for the feedback. Did the "system requirements" installation, and "Boost" and "Splat!" compilation as documented work for you without errors?
