### 2024-09-04 Swiss-Armed-Forces/openburst#1 @hasantahir
Hi @z-mathews

Boost was compiled whereas splat! was installed through `sudo apt install splat`. They seem to work fine.
