### 2024-09-05 9cfaea0
splat installation doc minor correction; buildBurst minor comment
