+++ docs/source/SPLAT_BURST_BOOST_README.md
@@ -7,8 +7,9 @@
 
 ### Convert SRTM files in batch to sdf files
 * as described in: [onetransistor.eu](https://www.onetransistor.eu/2016/07/splat-rf-compile-setup.html)
+* for converting the downladed hgt files to sdf files using the bash script as described in the link above, splat! installation is necessary (sudo apt install splat)
 
-### configure source/DEM/SPLAT_RADIOPROP/splatBurst.h:
+### configure radterrain/SPLAT_RADIOPROP/splatBurst.h:
  
 * replace SDF DEM files path in radterrain/SPLAT_RADIOPROP/splatBurst.h with the correct path to SDF files folder as in:
 ```
