### 2024-09-05 11baf58
minor changes to SPLAT_BURST_BOOST_README doc and to buildBurst comment
