### 2024-09-05 0af0671
minor changes to SPLAT_BURST_BOOST_README doc
