### 2024-09-05 98070fa
added some doc to install_readme
