### 2024-09-05 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir , following points:

- Splat! was extended and compiled as a shared library for openBURST with parallelizing the computations and for calling functions from Python. Therefore it is necessary to follow the Splat! compilation from the provided adapted source and making sure that shared libaries are created as documented. ` Note that `sudo apt install splat` is therefore not enough for installing Splat! for the purpose of openBURST (although it is necessary for conversion of the `hgt` height files to `sdf ` format, required by Splat!. This is documented as well.)

- I have updated the `pyproject.toml` file and tested the openBURST installation. Thank you for pointing this out.

- after installing openBURST, please do `pip show openburst` to check if openBURST is installed correctly.

- after installation is complete please also run the setup of Octave, postgreSQL DB (which includes running scripts to create the necessary tables) and port access as documented.

- after this the tests should run fine.

Please inform if you should encounter further issues. Thanks again.
