+++ docs/requirements.txt
@@ -17,6 +17,6 @@ simplekml>=1.3.6
 oct2py>=5.6.0
 geopandas>=0.13.2
 pandas>=2.0.1
-pyeval>=0.
+pyeval>=0.2
 scipy>=1.10.1
 pytest>=8.1.1
