### 2024-09-06 Swiss-Armed-Forces/openburst#1 @hasantahir
@z-mathews I am away from my Linux machine unfortunately until Monday. I will update this issue then. Thanks for patience.
