### 2024-09-17 Swiss-Armed-Forces/openburst#1 @danielskatz
@hasantahir - any news since 2 weeks ago?
