+++ pyproject.toml
@@ -30,6 +30,7 @@ dependencies = [
     'pyeval>=0.2',
     'scipy>=1.10.1 ',
     'pytest>=8.1.1',
+    'fiona>=1.10.1',
 ]
 
 license = {file = "LICENSE"}
