### 2024-09-25 #7052 @danielskatz
👋 @rvg296, @hasantahir, @thippesha-d - please remember that JOSS aims to have initial reviews completed in 2-4 weeks, and it's now been about 7. Can I ask you to please dedicate some time to continuing your reviews?
