### 2024-09-26 8d044c7
Update joss-paper-workflow.yml
