### 2024-09-27 Swiss-Armed-Forces/openburst#1 @hasantahir
Hi there,

Sorry for the delay in this. I have been trying to piece together a lot of packages. I have followed instructions to build `splat` and reached to the point of `mpi` installation when I encounter the following errors:

```shell
./build_mpi los_and_loss
Compiling BOOST MPI LOS AND LOSS... /usr/bin/ld: SONAME must not be empty string; ignored
/usr/bin/ld: .//libsplathd.so: undefined reference to `boost::python::instance_holder::allocate(_object*, unsigned long, unsigned long, unsigned long)'
collect2: error: ld returned 1 exit status
```

I also note that in the [instructions](https://openburst.readthedocs.io/en/latest/SPLAT_BURST_BOOST_README.html#build-with-mpi), I didn't have to include the flags `-lboost_python310` and `-lpython3.10` as they are already there.

This is how the initial lines of the shell script `build_mpi` looks like:

```shell
PYTHON_VERSION="3.10"
PYTHON_INC="/usr/include/python3.10"
BOOST_LIB_FILE="boost_python310" # this means
BOOST_INC="/usr/ include/boost"
BOOST_LIB_LOCATION="/usr/local/lib/"
BOOST_LIB_FILE="boost_python310" # this means
```
