### 2024-09-30 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir
Thank you for the message.  I am assuming you are using Python 3.10 and have edited the build files accordingly as in the installation instructions. Also, I am assuming that Boost was compiled and installed successfully according to the instructions.
Following two questions regarding the success of the installation steps:
- in the installation step called  "build splat-hd" when you run "./build splat" are the object files "splat" and "splat-hd" successfully generated?
- at the end of the next installation step called "build splat", were you able to run "./buildBurst splat_shared_lib" successfully? If successful, it should generate the shared library  file “libsplathd.so". This shared library is the one that is being used in the next step of installation with mpi.

Could you please let me know about the above two installation steps? Thanks.
