### 2024-09-30 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir
Thank you for the update. Compiling Splat! and creating the shared library seem to work. The warning:
`/usr/bin/ld:` SONAME must not be empty string; ignored`
can be ignored.
I see that your "build_mpi" script file above has a space in the variable definition "BOOST_INC" between "/usr/" and "include/boost". If that is the  case, please make sure, that the variable definition is correct, i.e. like this:
`BOOST_INC="/usr/include/boost"`
And then please try again compiling with MPI:
`./build_mpi los_and_loss`
And please let me know. Thanks.
