### 2024-09-30 Swiss-Armed-Forces/openburst#1 @hasantahir
Hi @z-mathews

Thanks for your reply. I checked the `build_mpi` shell script and there is no space in the path variable definition as you suggested:

```shell
PYTHON_VERSION="3.10"
PYTHON_INC="/usr/include/python3.10"
BOOST_LIB_FILE="boost_python310"
BOOST_INC="/usr/include/boost/"
BOOST_LIB_LOCATION="/usr/local/lib/"

```

The previous comment had a typo due to a screenshot tool.
