### 2024-09-30 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir
Thank you. Could you check if the Boost installation from source went well for you as in the [instructions. ](https://openburst.readthedocs.io/en/latest/INSTALL_BOOST.html)? Especially please check if your "/usr/lib" folder has the files:
`libboost_python310.a
libboost_python310.so
libboost_python310.so.1.85.0`

and your "/usr/local/include" has the boost header files.
