### 2024-09-30 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir please also make sure that when you compiled boost (before installing it locallly) that it was compiled with "python" and with "mpi" as shown in the boost compilation instructions [:](https://openburst.readthedocs.io/en/latest/INSTALL_BOOST.html)
`./bootstrap.sh --prefix=./TMP --with-libraries=python,mpi`
