### 2024-09-30 Swiss-Armed-Forces/openburst#1 @hasantahir
I confirm that the boost was built with python and mpi flags:
