### 2024-10-01 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir
thank you for the confirmations of the installation steps. To summarize:

- Boost compilation with Python & Boost and local installation seem to have worked
- Splat compilation and Splat shared library compilation seem to have worked (splat, splat-hd and libsplathd.so files created) -> I guess you have used the same linker flags in the script `buildBurst` as given in the repository, meaning linking with Boost and MPI.
- compilation of Splat parallelization source code with MPI (script `build_mpi`) gives you the linking error you mentioned:
`/usr/bin/ld: .//libsplathd.so: undefined reference to "boost::python::instance_holder::allocate(_object*, unsigned long, unsigned long, unsigned long)" collect2: error: ld returned 1 exit status`

From this, I think the order of the linker flags either in `buildBurst` or in `build_mpi` could be the cause. Could you please let me know the following, so that I can try to reproduce this error on a virtual machine:
- your Linux distribution name and version
- your Python version I understand is Python v3.10.12
- the version of Boost you downloaded and installed

Thank you.
