### 2024-10-02 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir thank you for the information. I could reproduce the linking error you had on a VM for your setup. The solution is to update the `build_mpi` acript with the compilation option `-fPIC` and the  linker option `-shared` to create the intended shared library. I have updated the `build_mpi` script in both `main` and `joss` branches accordingly. If you pull this version, you should be able to perform both compilations:

- `./build_mpi los_and_loss`
`./build_mpi radial_los_and_loss`

as in the documentation. They should create the files:

- `mpi_los_and_loss`
- `mpi_radial_los_and_loss`
in your directory.
