### 2024-10-03 Swiss-Armed-Forces/openburst#1 @hasantahir
Hi @z-mathews

Thanks. I have been successful in this and moving on, I am having difficulty in setting up the server through `pgadmin`. Specifically, it is unclear, what is meant by the [statement](https://openburst.readthedocs.io/en/latest/POSTGRESQL_README.html):

> - open pgadmin4 and then add server above with the username and password; leave host field and password fields empty, but use the username as your linux user name

I have inserted these settings:

but cant bring the server up and running.
