### 2024-10-04 0ea22cc
removed pyproject.toml formatting error for authors
