+++ pyproject.toml
@@ -2,15 +2,19 @@
 requires = ["setuptools", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
+[requires]
+python_version = "3.10"
+
 [project]
 name = "openburst"
-authors = [
-    {name = "Zenon Mathews", email = "zenon.mathews@vtg.admin.ch", name = "Romain Chessex", email = "romain.chessex@vtg.admin.ch"}
-]
 description = "A target and sensor simulator for air surveillance"
 version = "1.0"
 readme = "README.md"
-requires-python = "3.10"
+authors = [
+  {name = "Zenon Mathews", email = "zenon.mathews@vtg.admin.ch"},
+  {name = "Romain Chessex", email = "romain.chessex@vtg.admin.ch"},
+]
+
 dependencies = [
     'pyproj>=3.5.0',
     'psutil>=5.9.5',
