### 2024-10-04 4894d21
updated README.md for copying sample config file local config file
