### 2024-10-04 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir  great that your Spalt! and openBURST installations went well. To be able to use pgadmin to view your DB you should set "host name/ address" in the setting in your snapshot to "localhost". The username and Password should correspond to the ones you used to create the DB as in the [documentation](https://openburst.readthedocs.io/en/latest/POSTGRESQL_README.html).  readthedocs  is updated to make this clear. Thank you.
