### 2024-10-04 3b8c5d9
removed the -shared linker flag and the -fPIC compiler flag in buildBurst and build_mpi files as the linker error was due to wrong BOOST_LIB_LOCATION variable /usr/lib
