### 2024-10-04 8ddc1e0
extended readme for downloading DEM files
