### 2024-10-04 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir I did some more checks and noticed that the linker error you had with build_mpi could be because of the wrong variable definition in `buid_mpi` and `buildBurst` script files. Please make sure that the variable in those scripts is set correctly, i.e. BOOST_LIB_LOCATION="/usr/lib/", and not BOOST_LIB_LOCATION="/usr/local/lib/". At the end of the boost compilation the libraries were installed to "/usr/lib". So the problem could have been that the linker was not finding the boost_python libraries due to this error.  I have updated the repository and the readthedocs to denote this. Also in the repository I have removed again the -fPIC and -shared flags from build_mpi as I was able to compile without those flags for your setup. Could you please check that the above in your case and once more pull from the repository and do the following compilations once more:
`./buildBurst splat_shared_lib`
`./build_mpi los_and_loss`
`./build_mpi radial_los_and_loss`
Please inform if that worked. Thank you.
