### 2024-10-07 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir thank you for the message. I'm assuming you recompiled as in my previous [comment](https://github.com/Swiss-Armed-Forces/openburst/issues/1#issuecomment-2394568977) and as documented, and you also set up the  postgresql DB tables as in the documentation. In that case, could you please inform if you ran the tests as documented and if they were successful. Except the passive radar test, you should have all the other tests run successfully. The passive radar test will only be successful during real-time simulation of passive radar. If you can confirm the above it would mean that installation and testing were completed.
