### 2024-10-07 Swiss-Armed-Forces/openburst#1 @hasantahir
whilst setting up the postgresql schema and tables, I get the following errors:

```shell
psql: error: create_triggers.sql: No such file or directory
(openburst-venv) hasan@hasans:~/openburst$ cd openburst/db/
(openburst-venv) hasan@hasans:~/openburst/openburst/db$ psql red -f create_triggers.sql
CREATE FUNCTION
DROP TRIGGER
CREATE TRIGGER
DROP TRIGGER
CREATE TRIGGER
DROP TRIGGER
CREATE TRIGGER
DROP TRIGGER
CREATE TRIGGER
psql:create_triggers.sql:68: NOTICE:  relation "red_live.target" does not exist, skipping
DROP TRIGGER
psql:create_triggers.sql:69: ERROR:  relation "red_live.target" does not exist
DROP TRIGGER
CREATE TRIGGER
DROP TRIGGER
CREATE TRIGGER
```

I believe the DB server is running otherwise:

The only test information I was able to find in the documentation was the [command](https://openburst.readthedocs.io/en/latest/INSTALL_README.html#testing):
`pytest --pyargs openburst`

Running this throws errors:
```shell
=============================================== test session starts ===============================================
rootdir: /home/hasan/openburst
configfile: pyproject.toml
collected 2 items / 5 errors

===================================================== ERRORS ======================================================
_______________________________ ERROR collecting openburst/tests/db/test_db_conn.py _______________________________
ImportError while importing test module '/home/hasan/openburst/openburst/tests/db/test_db_conn.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst/tests/db/test_db_conn.py:6: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
_________________________ ERROR collecting openburst/tests/dem/test_get_terrain_height.py _________________________
ImportError while importing test module '/home/hasan/openburst/openburst/tests/dem/test_get_terrain_height.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst/tests/dem/test_get_terrain_height.py:8: in <module>
    from openburst.functions import basefunctions
openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
_______________________ ERROR collecting openburst/tests/openburst_splat/test_splat_load.py _______________________
ImportError while importing test module '/home/hasan/openburst/openburst/tests/openburst_splat/test_splat_load.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst/tests/openburst_splat/test_splat_load.py:8: in <module>
    from openburst.functions import basefunctions
openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
_________________ ERROR collecting openburst/tests/openburst_splat/test_splat_parallel_matrix.py __________________
ImportError while importing test module '/home/hasan/openburst/openburst/tests/openburst_splat/test_splat_parallel_matrix.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst/tests/openburst_splat/test_splat_parallel_matrix.py:11: in <module>
    from openburst.functions import basefunctions
openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
________________________ ERROR collecting openburst/tests/pcl/test_read_pcl_detections.py _________________________
ImportError while importing test module '/home/hasan/openburst/openburst/tests/pcl/test_read_pcl_detections.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst/tests/pcl/test_read_pcl_detections.py:7: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
============================================= short test summary info =============================================
ERROR openburst/tests/db/test_db_conn.py
ERROR openburst/tests/dem/test_get_terrain_height.py
ERROR openburst/tests/openburst_splat/test_splat_load.py
ERROR openburst/tests/openburst_splat/test_splat_parallel_matrix.py
ERROR openburst/tests/pcl/test_read_pcl_detections.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 5 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
================================================ 5 errors in 0.30s ================================================
```
