+++ docs/source/RUNNING_OPENBURST.md
@@ -1,0 +1,19 @@
+ 
+
+## Running openBURST
+
+This assumes that openBURST was installed correctly as documented. 
+
+### Starting the server modules
+
+* run modules_manager.py from folder openburst/ and start the necessary modules by pressing "Start" or "Stop". Show the log files for the server using "Show Log". Each log window can be closed by simply pressing "q" after clicking on the log window. The current run status of all server modules can be checked by pressing "Check", when all currently running servers will be indicated as "running". Stop all server modules before quitting by pressing "Quit".  
+
+
+![Servers](./images/servers_gui.png?raw=true "Server Modules GUI")
+
+### Starting the web browser client
+
+* then open a browser and type in the url: http://localhost:8888
+
+
+
