### 2024-10-07 3dba56c
extended readme_rtd
