+++ docs/source/RUNNING_OPENBURST.md
@@ -6,27 +6,27 @@ This assumes that openBURST was installed correctly as documented.
 
 ### Starting the server modules
 
-* run modules_manager.py from folder openburst/ and start the necessary modules by pressing "Start" or "Stop". Show the log files for the server using "Show Log". Each log window can be closed by simply pressing "q" after clicking on the log window. The current run status of all server modules can be checked by pressing "Check", when all currently running servers will be indicated as "running". Stop all server modules before quitting by pressing "Quit".  
+Run "python modules_manager.py" (make sure to use the correct python version) from folder openburst/ and start the necessary modules by pressing "Start" (use "Stop" to stop the modules later). Show the log files for the server using "Show Log". Each log window can be closed by simply pressing "q" after clicking on the log window. The current run status of all server modules can be checked by pressing "Check", when all currently running servers will be indicated as "running". Stop all server modules by pressing "Stop" for each server module before quitting by pressing "Quit".  
 
 ![Alt text](./images/servers_gui.png "Server Modules GUI")
 
 ### Starting the web browser client
 
-Open a browser and type in the url: http://localhost:8888
-You should get the following login page. Log in using the role "theater" by removing "/airpic" from the value in the second filed (the other role "airpic" is for viewing an airpicture). Then press the "Launch" button. 
+Open a browser and type in the url: "http://localhost:8888".
+You should get the following login page. Log in using the role "theater" by removing "/airpic" from the value in the second field (the other role "airpic" is for viewing an air picture). Then press the "Launch" button. 
 
 ![Alt text](./images/openburst_logging.png "Log in")
 
 
 ### Elevation information
 
-after logging in you should see a webpage with an openstreetmap (internet access is necessary). Double left click on the map to place the POI(Point of Intrerest)  icon on the map. Left click with the mouse on the map and you should be able see a pop up window with the lat, lon, masl (altitude in meters above sea level) and the distance to the previously placed POI. Note that you should have downloaded the DEM (digital elevation model) files for the area aou are interested in.
+After logging in you should see a webpage with an openstreetmap (internet access is necessary). Double left click on the map to place the POI(Point of Intrerest) icon on the map. Then left click with the mouse on another position on the map and you should be able see a pop up window with the lat, lon, masl (altitude in meters above sea level) and the distance to the previously placed POI. Note that you should have downloaded the DEM (digital elevation model) files for the area you are interested in.
 
 ![Alt text](./images/masl.png "Masl")
 
 ### Terrain backdrop
 
-To get the backdrop of certain point on map, place the POI on the map and press the "TERR button at the bottom of the page. Be aware that all terrain computations needs downloading the necessary DEM files as documented. In the "TERR" tab on the right you should get the terrain backdrop for a 10x10 area with the POI in the center. See below the example for POI at the Matterhorn peak in the southern Swiss alps. 
+To get the backdrop of a certain point on the map, place the POI on the map and press the "TERR" button at the bottom of the page. Be aware that all terrain computations needs downloading the necessary DEM files as documented. In the "TERR" tab on the right you should now get an interactive terrain backdrop plot for a 10x10 area with the POI in the center. See below the example for POI at the Matterhorn peak in the southern Swiss alps. This feature can be used to visually inspect a potential sensor location.       
 
 ![Alt text](./images/terrain_matterhorn.png "terrain")
 
@@ -36,17 +36,17 @@ To do a sample calculation of propagation losses place the POI on the map (doubl
 
 ![Alt text](./images/poi_and_point.png "poi and point")
 
-Now move to the tab "Radio propagation" and press the button "Detailed analysis". You should now get the propagation model results for a given frequency between the two above placed points on the map. 
[…]
