### 2024-10-08 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir thank you for the message. Could you please confirm the following?

1. Did the step "openburst installation" in the [documentation](https://openburst.readthedocs.io/en/latest/INSTALL_README.html) work without errors? Ie. you could pip install openburst.
2. In the "Setup" section of the installation [documentation ](https://openburst.readthedocs.io/en/latest/INSTALL_README.html) did you do the "Octave", "postgreSQL" and "Run configuration" configurations?
3. Please make sure that in the step "Run configuration" above is performed before running the tests. I.e. copy the "constants/openburst_config_local.py" to "constants/openburst_config.py" and set the parameters indicated in the file for your setup.
4. Please then run the pytest again as above.
Thank you.
