### 2024-10-08 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir once you are have the previous configuration steps performed and have run the pytests, you can check the core functionalities of openburst as documented [here](https://openburst.readthedocs.io/en/latest/RUNNING_OPENBURST.html).
