### 2024-10-08 #7052 @hasantahir
Hi @danielskatz

I have spent a considerable amount of time building the tool and I have been in touch with the developers to address [issues](https://github.com/Swiss-Armed-Forces/openburst/issues/1#issuecomment-2397258311). I am close to running the test cases after which I will update this review. Thanks for the patience - it has been a long process.
