### 2024-10-08 Swiss-Armed-Forces/openburst#1 @hasantahir
@z-mathews thanks for your message. I believe openburst, Octave and the database have been set up:

```shell
(openburst-venv) hasan@hasans:~/openburst/openburst/constants$ pip show openburst
Name: openburst
Version: 1.0
Summary: A target and sensor simulator for air surveillance
Home-page:
Author:
Author-email:
License: GNU GENERAL PUBLIC LICENSE
                               Version 3, 29 June 2007
```

```shell
(openburst-venv) hasan@hasans:~/openburst/openburst/constants$ which octave
/usr/bin/octave
```

```shell
hasan@hasans:/etc/postgresql/14/main$ psql -d red
psql (14.13 (Ubuntu 14.13-0ubuntu0.22.04.1))
Type "help" for help.

red=#
```

This is evident by the dashboard activity:

The files are also modified as per [PostgreSQL readme](https://openburst.readthedocs.io/en/latest/POSTGRESQL_README.html). I also receive these errors:
```shell
