### 2024-10-08 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir thank you for the message, it means that the IP address in the script "write_table_entries.sql" you should replace the default IP address "'192.111.1.177'" with your machines IP address before running it.
