### 2024-10-08 Swiss-Armed-Forces/openburst#1 @hasantahir
thanks, in the `create_tables.sql` should the lines be like this:
```shell
ALTER TABLE blue_live.pcl_detection
   OWNER TO hasan;
```
