### 2024-10-08 Swiss-Armed-Forces/openburst#1 @hasantahir
The only error in this process that I get is this:

```
hasan@hasans:~/openburst/openburst/db$ psql red -f create_triggers.sql
CREATE FUNCTION
DROP TRIGGER
CREATE TRIGGER
DROP TRIGGER
CREATE TRIGGER
DROP TRIGGER
CREATE TRIGGER
DROP TRIGGER
CREATE TRIGGER
psql:create_triggers.sql:68: NOTICE:  relation "red_live.target" does not exist, skipping
DROP TRIGGER
psql:create_triggers.sql:69: ERROR:  relation "red_live.target" does not exist
DROP TRIGGER
CREATE TRIGGER
DROP TRIGGER
CREATE TRIGGER
```
