### 2024-10-08 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir yes the username should be the username of the DB username you had set before while creating the postgresql DB. The error for the "red" table can be ignored as extending openBURST with "red" tables for "blue" Vs "red" gaming is currently a growth potential.
