### 2024-10-09 Swiss-Armed-Forces/openburst#1 @hasantahir
I ran the test again, and it seems there Python is not be recognised:
```shell
(openburst-venv) hasan@hasans:~/openburst$ pytest --pyargs openburst
============================================== test session starts ===============================================
rootdir: /home/hasan/openburst
configfile: pyproject.toml
collected 2 items / 5 errors

===================================================== ERRORS =====================================================
______________________________ ERROR collecting openburst/tests/db/test_db_conn.py _______________________________
ImportError while importing test module '/home/hasan/openburst/openburst/tests/db/test_db_conn.py'.
Traceback:
/usr/lib/python3.10/importlib/__init__.py:126: in import_module
openburst/tests/db/test_db_conn.py:6: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
E   ImportError: cannot import name 'openburst_config' from 'openburst.constants' (unknown location)
```

any suggestions on this? I believe we are close :-).
