### 2024-10-09 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir I too believe we are close :-). Did you create a file name "openburst_config.py" in the constants folder by copying the file named "openburst_config_local.py" and configuring the parameters as commented in the file?
