### 2024-10-09 Swiss-Armed-Forces/openburst#1 @hasantahir
Thanks @z-mathews, I was not creating a new file but had modified the `openburst_config_local.py` earlier on. Now the contents of the `openburst_config.py` file look like this:
```shell
#---- set python version
PYTHON_VERSION = "python3.10"

# directory where KML files are saved on the machine of the geoplot server
KML_FILE_PATH = "/home/red3/Documents/OPENBURST/KML/"

# postgresql DB parameters
BURST_DB_NAME = "red"
BURST_DB_SERVER_IP = "XXX.XXX.XXX.XXX" # replaced the actual IP
BURST_DB_SERVER_USERNAME = "hasan"
BURST_DB_SERVER_PASSWORD = "XXX" # replaced the password
```

Now I get the following errors:
```shell
(openburst-venv) hasan@hasans:~/openburst$ pytest --pyargs openburst
============================================== test session starts ===============================================
rootdir: /home/hasan/openburst
configfile: pyproject.toml
collected 2 items / 5 errors

===================================================== ERRORS =====================================================
______________________________ ERROR collecting openburst/tests/db/test_db_conn.py _______________________________
openburst/tests/db/test_db_conn.py:6: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
openburst/constants/openburst_config.py:27: in <module>
    ROOT_DIR = os.path.dirname(os.path.abspath(inspect.getsourcefile(openburst)))
/usr/lib/python3.10/inspect.py:817: in getsourcefile
    filename = getfile(object)
/usr/lib/python3.10/inspect.py:778: in getfile
    raise TypeError('{!r} is a built-in module'.format(object))
E   TypeError: <module 'openburst' (<_frozen_importlib_external._NamespaceLoader object at 0x754f0a1d2c20>)> is a built-in module
________________________ ERROR collecting openburst/tests/dem/test_get_terrain_height.py _________________________
openburst/tests/dem/test_get_terrain_height.py:8: in <module>
    from openburst.functions import basefunctions
openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
openburst/constants/openburst_config.py:27: in <module>
    ROOT_DIR = os.path.dirname(os.path.abspath(inspect.getsourcefile(openburst)))
/usr/lib/python3.10/inspect.py:817: in getsourcefile
    filename = getfile(object)
/usr/lib/python3.10/inspect.py:778: in getfile
    raise TypeError('{!r} is a built-in module'.format(object))
E   TypeError: <module 'openburst' (<_frozen_importlib_external._NamespaceLoader object at 0x754f0a1d2c20>)> is a built-in module
______________________ ERROR collecting openburst/tests/openburst_splat/test_splat_load.py _______________________
openburst/tests/openburst_splat/test_splat_load.py:8: in <module>
    from openburst.functions import basefunctions
openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
openburst/constants/openburst_config.py:27: in <module>
    ROOT_DIR = os.path.dirname(os.path.abspath(inspect.getsourcefile(openburst)))
/usr/lib/python3.10/inspect.py:817: in getsourcefile
    filename = getfile(object)
/usr/lib/python3.10/inspect.py:778: in getfile
    raise TypeError('{!r} is a built-in module'.format(object))
E   TypeError: <module 'openburst' (<_frozen_importlib_external._NamespaceLoader object at 0x754f0a1d2c20>)> is a built-in module
_________________ ERROR collecting openburst/tests/openburst_splat/test_splat_parallel_matrix.py _________________
openburst/tests/openburst_splat/test_splat_parallel_matrix.py:11: in <module>
    from openburst.functions import basefunctions
openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
openburst/constants/openburst_config.py:27: in <module>
    ROOT_DIR = os.path.dirname(os.path.abspath(inspect.getsourcefile(openburst)))
/usr/lib/python3.10/inspect.py:817: in getsourcefile
    filename = getfile(object)
/usr/lib/python3.10/inspect.py:778: in getfile
    raise TypeError('{!r} is a built-in module'.format(object))
E   TypeError: <module 'openburst' (<_frozen_importlib_external._NamespaceLoader object at 0x754f0a1d2c20>)> is a built-in module
________________________ ERROR collecting openburst/tests/pcl/test_read_pcl_detections.py ________________________
openburst/tests/pcl/test_read_pcl_detections.py:7: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
openburst/constants/openburst_config.py:27: in <module>
    ROOT_DIR = os.path.dirname(os.path.abspath(inspect.getsourcefile(openburst)))
/usr/lib/python3.10/inspect.py:817: in getsourcefile
    filename = getfile(object)
/usr/lib/python3.10/inspect.py:778: in getfile
    raise TypeError('{!r} is a built-in module'.format(object))
E   TypeError: <module 'openburst' (<_frozen_importlib_external._NamespaceLoader object at 0x754f0a1d2c20>)> is a built-in module
============================================ short test summary info =============================================
ERROR openburst/tests/db/test_db_conn.py - TypeError: <module 'openburst' (<_frozen_importlib_external._NamespaceLoader object at 0x754f0a1d2c20>)> is a...
ERROR openburst/tests/dem/test_get_terrain_height.py - TypeError: <module 'openburst' (<_frozen_importlib_external._NamespaceLoader object at 0x754f0a1d2c20>)> is a...
ERROR openburst/tests/openburst_splat/test_splat_load.py - TypeError: <module 'openburst' (<_frozen_importlib_external._NamespaceLoader object at 0x754f0a1d2c20>)> is a...
ERROR openburst/tests/openburst_splat/test_splat_parallel_matrix.py - TypeError: <module 'openburst' (<_frozen_importlib_external._NamespaceLoader object at 0x754f0a1d2c20>)> is a...
ERROR openburst/tests/pcl/test_read_pcl_detections.py - TypeError: <module 'openburst' (<_frozen_importlib_external._NamespaceLoader object at 0x754f0a1d2c20>)> is a...
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 5 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
=============================================== 5 errors in 0.56s ================================================
```

`pip show openburst` indeed gives me the expected information:
```shell
(openburst-venv) hasan@hasans:~/openburst/openburst/constants$ pip show openburst
Name: openburst
Version: 1.0
Summary: A target and sensor simulator for air surveillance
Home-page:
Author:
Author-email:
```

Any suggestions?
