### 2024-10-10 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir I cannot reproduce the error on the VM I have for your setup. I am guessing, it might have to do with the recompilations of splat and shared libraries etc. that were done. I suggest to do the following to have it clean:

1. uninstall openburst by  `pip uninstall openburst`
2. make sure you have the latest version of the following build files as currently in the repo in the directory "openburst/radterrrain/SPLAT_RADIOPROP" (you can git pull them):
`build`
`buildBurst`
`build_mpi`

3.  check once more the following variables are set correctly in your build files: PYTHON_VERSION, PYTHON_INC, BOOST_INC, BOOST_LIB_LOCATION, BOOST_LIB_FILE
4. build again everthing in the directory "openburst/radterrrain/SPLAT_RADIOPROP" by the following commands:
`./build splat`
`./buildBurst splat_shared_lib`
`./build_mpi los_and_loss`
`./build_mpi radial_los_and_loss`

5. check that the following object files were newly created: "splat", "splat-hd", "libsplathd.so", "mpi_los_and_loss" and "mpi_radial_los_and_loss"
6. now pip install openburst again as shown [here ](https://openburst.readthedocs.io/en/latest/INSTALL_README.html)
7. after this try the tests again
