### 2024-10-10 Swiss-Armed-Forces/openburst#1 @hasantahir
Unfortunately, I am still getting the same error after unistalling openburst, fresh pull, building and reinstalling openburst.
