### 2024-10-10 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir not sure if invoking pytest from your virtual-env plays a role here for your setup. can you please try the following:
`python -m pytest --pyargs openburst`
