### 2024-10-10 Swiss-Armed-Forces/openburst#1 @hasantahir
Unfortunately its the same in or outside the virtual environment.
