### 2024-10-11 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir thank you. as `inspect.getsourcefile` seems to be throwing the error under pytest, could you please replace the following line in your "constants/openburst_config.py"
`ROOT_DIR = os.path.dirname(os.path.abspath(inspect.getsourcefile(openburst)))`
with
`ROOT_DIR = os.path.dirname(openburst.__file__)`
save the file and then run the tests again?
