### 2024-10-11 Swiss-Armed-Forces/openburst#1 @hasantahir
After making the path change, I see the following error messages:
```shell
(openburst-venv) hasan@hasans:~/openburst$ pytest --pyargs openburst
============================================================ test session starts ============================================================
rootdir: /home/hasan/openburst
configfile: pyproject.toml
collected 2 items / 5 errors

================================================================== ERRORS ===================================================================
____________________________________________ ERROR collecting openburst/tests/db/test_db_conn.py ____________________________________________
openburst/tests/db/test_db_conn.py:6: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
openburst/constants/openburst_config.py:28: in <module>
    ROOT_DIR = os.path.dirname(openburst.__file__)
/usr/lib/python3.10/posixpath.py:152: in dirname
    p = os.fspath(p)
E   TypeError: expected str, bytes or os.PathLike object, not NoneType
______________________________________ ERROR collecting openburst/tests/dem/test_get_terrain_height.py ______________________________________
openburst/tests/dem/test_get_terrain_height.py:8: in <module>
    from openburst.functions import basefunctions
openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
openburst/constants/openburst_config.py:28: in <module>
    ROOT_DIR = os.path.dirname(openburst.__file__)
/usr/lib/python3.10/posixpath.py:152: in dirname
    p = os.fspath(p)
E   TypeError: expected str, bytes or os.PathLike object, not NoneType
____________________________________ ERROR collecting openburst/tests/openburst_splat/test_splat_load.py ____________________________________
openburst/tests/openburst_splat/test_splat_load.py:8: in <module>
    from openburst.functions import basefunctions
openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
openburst/constants/openburst_config.py:28: in <module>
    ROOT_DIR = os.path.dirname(openburst.__file__)
/usr/lib/python3.10/posixpath.py:152: in dirname
    p = os.fspath(p)
E   TypeError: expected str, bytes or os.PathLike object, not NoneType
______________________________ ERROR collecting openburst/tests/openburst_splat/test_splat_parallel_matrix.py _______________________________
openburst/tests/openburst_splat/test_splat_parallel_matrix.py:11: in <module>
    from openburst.functions import basefunctions
openburst/functions/basefunctions.py:18: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
openburst/constants/openburst_config.py:28: in <module>
    ROOT_DIR = os.path.dirname(openburst.__file__)
/usr/lib/python3.10/posixpath.py:152: in dirname
    p = os.fspath(p)
E   TypeError: expected str, bytes or os.PathLike object, not NoneType
_____________________________________ ERROR collecting openburst/tests/pcl/test_read_pcl_detections.py ______________________________________
openburst/tests/pcl/test_read_pcl_detections.py:7: in <module>
    from openburst.functions import dbfunctions
openburst/functions/dbfunctions.py:16: in <module>
    from openburst.constants import openburst_config
openburst/constants/openburst_config.py:28: in <module>
    ROOT_DIR = os.path.dirname(openburst.__file__)
/usr/lib/python3.10/posixpath.py:152: in dirname
    p = os.fspath(p)
E   TypeError: expected str, bytes or os.PathLike object, not NoneType
========================================================== short test summary info ==========================================================
ERROR openburst/tests/db/test_db_conn.py - TypeError: expected str, bytes or os.PathLike object, not NoneType
ERROR openburst/tests/dem/test_get_terrain_height.py - TypeError: expected str, bytes or os.PathLike object, not NoneType
ERROR openburst/tests/openburst_splat/test_splat_load.py - TypeError: expected str, bytes or os.PathLike object, not NoneType
ERROR openburst/tests/openburst_splat/test_splat_parallel_matrix.py - TypeError: expected str, bytes or os.PathLike object, not NoneType
ERROR openburst/tests/pcl/test_read_pcl_detections.py - TypeError: expected str, bytes or os.PathLike object, not NoneType
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 5 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
============================================================= 5 errors in 0.36s =============================================================
```
Thanks. This may be suggesting some directories are empty which they should not be, perhaps. Thoughts?
