### 2024-10-11 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir in order to understand what is going on, could please start run python in your console and do the following:

`import os
os.path.dirname(os.__file__)
`
and see what output you get. Then also see the output for:

`
import openburst
os.path.dirname(openburst.__file__)
`
