### 2024-10-11 Swiss-Armed-Forces/openburst#1 @hasantahir
```shell
>>> import os
>>> os.path.dirname(os.__file__)
'/usr/lib/python3.10'

>>> import openburst
>>> os.path.dirname(openburst.__file__)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/usr/lib/python3.10/posixpath.py", line 152, in dirname
    p = os.fspath(p)
TypeError: expected str, bytes or os.PathLike object, not NoneType
```

thanks, I am getting an error in I guess locating the whereabouts of the openburst path. I will try thay.
