### 2024-10-11 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir also, check that you still have "__init.py__" in your openburst root folder.
