### 2024-10-11 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir the second test error is because in the test file  "tests/dem/test_get_terrain_height.py" you should use the lat and lon values for a position where you have the DEM files. In the script a default value in Switzerland is used.
`lat = 47.0`
`lon = 8.1`
