### 2024-10-11 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir thanks. Yes, it is because openburst/splat! expects "hd" DEM files as [documented](https://openburst.readthedocs.io/en/latest/SPLAT_BURST_BOOST_README.html). You should use `srtm2sdf-hd` as documented to get the sdf DEM files with the "-hd" extension. Please convert the hgt files once more accordingly and try the test. It should work.
