### 2024-10-11 9c63004
updated INSTALL_README with a check for openburst package finding; uipdated openburst_config_local.py to not use inspect to find the ROOT_DIR
