### 2024-10-11 bf5702e
removed unnecessary pip uninstall openburst
