### 2024-10-11 Swiss-Armed-Forces/openburst#1 @hasantahir
I tried changing the coordinates of the `test_get_terrain_height.py` according to the tile that I downloaded but I am getting the same error. I have performed the `srtm2sdf-hd` conversion as well.
