### 2024-10-11 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir do you have the newly converted "XYZ-hd.sdf" files in the configured DEM folder "/home/hasan/Downloads/O29"?
