### 2024-10-11 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir could you make sure that you do NOT have the forward slash “/” at the end of the DEM pathname as above
 in the constants/splatconsants.py . E.g.

`DEM_FILES_PATH = "~/Downloads/SRTM3_Eurasia_Data/SDF_Files"`
