### 2024-10-11 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir in the test "test_get_terrain_height" the following assert:
 `assert alt==916`
will obviously fail for your location. The unit test is to just see if the terrain height for the given location is fetched correctly.
