### 2024-10-11 Swiss-Armed-Forces/openburst#1 @hasantahir
Hi,
Yes. please have a look:
```shell
hasan@hasans:~/Downloads/O29$ ls
56:57:6:7-hd.sdf  56:57:7:8.sdf     57:58:7:8-hd.sdf  57:58:8:9.sdf     58:59:7:8-hd.sdf  59:60:6:7.sdf  N57W007.hgt  N58W007.hgt  script.sh
56:57:6:7.sdf     57:58:6:7-hd.sdf  57:58:7:8.sdf     58:59:6:7-hd.sdf  58:59:7:8.sdf     N56W007.hgt    N57W008.hgt  N58W008.hgt
56:57:7:8-hd.sdf  57:58:6:7.sdf     57:58:8:9-hd.sdf  58:59:6:7.sdf     59:60:6:7-hd.sdf  N56W008.hgt    N57W009.hgt  N59W007.hgt
hasan@hasans:~/Downloads/O29$ pwd
/home/hasan/Downloads/O29
```
My DEM file path is:

The test log is:

```shell
(openburst-venv2) hasan@hasans:~/openburst/openburst/radterrain/SPLAT_RADIOPROP$ pytest --pyargs openburst
============================================================ test session starts ============================================================
rootdir: /home/hasan/openburst
configfile: pyproject.toml
collected 9 items

../../tests/db/test_db_conn.py .                                                                                                      [ 11%]
../../tests/dem/test_dem_folder.py F                                                                                                  [ 22%]
../../tests/dem/test_get_terrain_height.py F                                                                                          [ 33%]
../../tests/openburst_splat/test_splat_load.py ...                                                                                    [ 66%]
../../tests/openburst_splat/test_splat_parallel_matrix.py .                                                                           [ 77%]
../../tests/pcl/test_read_pcl_detections.py .                                                                                         [ 88%]
../../tests/rad/test_rad.py .                                                                                                         [100%]

================================================================= FAILURES ==================================================================
____________________________________________________________ test_get_dem_files _____________________________________________________________

    def test_get_dem_files():
        try:
            for fname in os.listdir(splatconstants.DEM_FILES_PATH):
              if not fname.endswith('.sdf'):
>               pytest.fail("not .sdf file found in folder..")
E               Failed: not .sdf file found in folder..

../../tests/dem/test_dem_folder.py:14: Failed
__________________________________________________________ test_get_terrain_height __________________________________________________________

    def test_get_terrain_height():
        try:
            lat = 57.41
            lon = -6.19
            alt = geofunctions.get_terrain_height(splat,lat, lon)
>           assert alt==916 # mASL = 916 at 47.0, 8.1
E           assert 0 == 916

../../tests/dem/test_get_terrain_height.py:20: AssertionError

    def test_get_terrain_height():
        try:
            lat = 57.41
            lon = -6.19
            alt = geofunctions.get_terrain_height(splat,lat, lon)
            assert alt==916 # mASL = 916 at 47.0, 8.1
        except Exception: # pylint: disable=bare-except
>           pytest.fail("Could not read terrain height ..")
E           Failed: Could not read terrain height ..

../../tests/dem/test_get_terrain_height.py:22: Failed
========================================================== short test summary info ==========================================================
FAILED ../../tests/dem/test_dem_folder.py::test_get_dem_files - Failed: not .sdf file found in folder..
FAILED ../../tests/dem/test_get_terrain_height.py::test_get_terrain_height - Failed: Could not read terrain height ..
======================================================== 2 failed, 7 passed in 7.40s ========================================================
 [INFO] Remove shared memory segment { name = dem_attrs_shared_memory_segment }
 [INFO] Remove shared memory segment { name = dem_shared_memory_segment }
```
