### 2024-10-11 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir I see that you should remove the forward slash "/" in the DEM_FILES_PATH. See [comment](https://github.com/Swiss-Armed-Forces/openburst/issues/1#issuecomment-2407712214).
