### 2024-10-14 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir also, please make sure that the DEM file  folder configuration in the file  "/DEM/SPLAT_RADIOPROP/splatBurst.h" is also done correctly and [as documented ](https://openburst.readthedocs.io/en/latest/SPLAT_BURST_BOOST_README.html) (i.e. with the forward slash "/" in this case) . If you need to make this change "splatBurst.h" file, please do recompile in directory "radterrain/SPLAT_RADIOPROP" as [documented](https://openburst.readthedocs.io/en/latest/SPLAT_BURST_BOOST_README.html) by:
`./build splat`
`./buildBurst splat_shared_lib`
`./build_mpi los_and_loss`
`./build_mpi radial_los_and_loss`
