### 2024-10-14 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir also, the test "test_get_dem_files" as it is now will fail, if non ".sdf" files are found in your DEM folder. So please make sure that your ".hgt" files are not in the above configured DEM folder. Thanks.
