### 2024-10-14 4e4f773
better documentation for making sure the DEM file folder contains just .sdf files
