### 2024-10-14 02183f8
better documentation for making sure the DEM file folder contains just .sdf files
