### 2024-10-18 Swiss-Armed-Forces/openburst#1 @z-mathews
Hi @hasantahir thank you for the message. By tutorials you mean the testing of core functionalities described [here](https://openburst.readthedocs.io/en/latest/RUNNING_OPENBURST.html)? If yes, you can follow the steps described their actually for any part of the world, i.e. for the region you have downloaded the DEM files for. If you want the exact region described there, then you need to download the L32 tile as documented from [here](https://www.viewfinderpanoramas.org/Coverage%20map%20viewfinderpanoramas_org3.htm).

bye the way, did the tests run fine in that case?
