### 2024-10-21 #7052 @danielskatz
👋 @thippesha-d - How is your review coming along? I don't see many filled checkboxes in your [list](https://github.com/openjournals/joss-reviews/issues/7052#issuecomment-2273710088) yet.
