### 2024-10-21 Swiss-Armed-Forces/openburst#1 @hasantahir
Hi @z-mathews

I have not reached the point of testing the functionalities of Openburst yet as the test runs still fail. I have tried removing the `/` as suggested, and recompiled but still end with failed tests:

```shell
(openburst-venv2) hasan@hasans:~/openburst$ pytest --pyargs openburst
============================================================ test session starts ============================================================
rootdir: /home/hasan/openburst
configfile: pyproject.toml
collected 9 items

openburst/tests/db/test_db_conn.py .                                                                                                  [ 11%]
openburst/tests/dem/test_dem_folder.py F                                                                                              [ 22%]
openburst/tests/dem/test_get_terrain_height.py F                                                                                      [ 33%]
openburst/tests/openburst_splat/test_splat_load.py ...                                                                                [ 66%]
openburst/tests/openburst_splat/test_splat_parallel_matrix.py .                                                                       [ 77%]
openburst/tests/pcl/test_read_pcl_detections.py .                                                                                     [ 88%]
openburst/tests/rad/test_rad.py .                                                                                                     [100%]

================================================================= FAILURES ==================================================================
____________________________________________________________ test_get_dem_files _____________________________________________________________

    def test_get_dem_files():
        try:
            for fname in os.listdir(splatconstants.DEM_FILES_PATH):
              if not fname.endswith('.sdf'):
>               pytest.fail("not .sdf file found in folder..")
E               Failed: not .sdf file found in folder..

openburst/tests/dem/test_dem_folder.py:14: Failed
__________________________________________________________ test_get_terrain_height __________________________________________________________

    def test_get_terrain_height():
        try:
            lat = 57.41
            lon = -6.19
            alt = geofunctions.get_terrain_height(splat,lat, lon)
>           assert alt==916 # mASL = 916 at 47.0, 8.1
E           assert -999 == 916

openburst/tests/dem/test_get_terrain_height.py:20: AssertionError

    def test_get_terrain_height():
        try:
            lat = 57.41
            lon = -6.19
            alt = geofunctions.get_terrain_height(splat,lat, lon)
            assert alt==916 # mASL = 916 at 47.0, 8.1
        except Exception: # pylint: disable=bare-except
>           pytest.fail("Could not read terrain height ..")
E           Failed: Could not read terrain height ..

openburst/tests/dem/test_get_terrain_height.py:22: Failed
----------------------------------------------------------- Captured stdout call ------------------------------------------------------------
:::::::: total DEM size = 3732480000
ok2: going to load topo
 -------------------> sdf_path /home/hasan/Downloads/O29: LoadSDF, name = 57:58:6:7-hd, ret_val = -1
LoadSDF_SDF failed...assuming water level 0masl
Region  "57:58:6:7-hd" assumed as sea-level into page 1...
========================================================== short test summary info ==========================================================
FAILED openburst/tests/dem/test_dem_folder.py::test_get_dem_files - Failed: not .sdf file found in folder..
FAILED openburst/tests/dem/test_get_terrain_height.py::test_get_terrain_height - Failed: Could not read terrain height ..
======================================================== 2 failed, 7 passed in 7.30s ========================================================
 [INFO] Remove shared memory segment { name = dem_attrs_shared_memory_segment }
 [INFO] Remove shared memory segment { name = dem_shared_memory_segment }
 ```
 My feeling is I am not able to pick up SDF files hence requesting you to supply ones through which I can pass the tests.
