### 2024-10-22 ccac146
updated dem tests
