+++ openburst/tests/dem/try_get_terrain_height.py
@@ -4,7 +4,7 @@ tests terrain height can be computed for a given lat/lon position
 
 """
 
-import pytest
+#import pytest
 from openburst.functions import basefunctions
 from openburst.functions import geofunctions
 basefunctions.set_openburst_system_path()
@@ -12,13 +12,22 @@ basefunctions.set_openburst_linked_lib_path()
 import libsplathd as splat
 
 
-def test_get_terrain_height():
+def try_get_terrain_height():
     try:
-        #lat = 47.0
-        #lon = 8.1
-        lat = 57.41
-        lon = -6.19
+        lat = 47.0
+        lon = 8.1
+        #lat = 57.41
+        #lon = -6.19
         alt = geofunctions.get_terrain_height(splat,lat, lon)
-        assert alt > -160 # we make sure that there is some masl read. assuming -160masl the lowest location below sea level
+        print("alt = ", alt)
+        #assert alt==916 # mASL = 916 at 47.0, 8.1
     except Exception: # pylint: disable=bare-except
-        pytest.fail("Could not read terrain height ..")
+        print("Could not read terrain height ..")
+
+
+def main():
+    try_get_terrain_height()
+
+
+if __name__ == "__main__":
+    main()
