### 2024-10-22 fafd03f
updated get terrrain height test
