### 2024-10-22 4ca3131
added a simple routine to get terrain height in tests/dem
