### 2024-10-22 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir thank you for the information. The first test error says that at least one non .sdf file was found in the DEM folder. This was used to make sure that the DEM folder has only .sdf files, which is actually not necessary, as the folder can also contain other files. So, I have change  the test to check if atleast one .sdf file is in the DEM folder, which is a necessity. The second test error is because the height check if made for the location lat/lon values (47.0, 8.1), which is 916masl. I have also changed this test to check if the read height is above -150masl, which is assumed to be the lowest masl below sea level on land. I downloaded the the .hgt tile for your point of interest (57.41, -6.19), and after conversion to .sdf that would be the sdf file "57:58:6:7-hd.sdf". And the test works for this. Also, I have added a new python script "try_get_terrain_height.py" in the tests/dem folder, which you can use to simply get the masl for a given height. For the lat,lon values 57.41,-6.19, running the above script by `python3 try_get_terrain_height.py` should print the value "alt=0". Could you please pull from the repository these changes and try the test once more?   I am assuming that you have the .sdf file  "57:58:6:7-hd.sdf" in your configfured DEM repository.
