### 2024-10-23 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir could you please paste here the following two things:

1. the result of "ls -la" in your DEM directory, so that we make sure you have the .sdf file with the correct size.
2. line 247 of the header file "splatBurst.h" in folder "radterrain/SPLAT_RADIOPROP"? That is the line that defines the DEM folder for splat and should look something like this:
`char burst_dem_path[] = "/home/user/Downloads/SDF_Files/";`
