### 2024-10-23 Swiss-Armed-Forces/openburst#1 @hasantahir
Hi @z-mathews

I have downloaded the L32 tile, pulled the latest updates and recompiled. A positive change, only one test fails.

Please find the outputs below:

```shell
(openburst-venv2) hasan@hasans:~/openburst/openburst/tests/dem$ python3 try_get_terrain_height.py
:::::::: total DEM size = 3732480000
ok2: going to load topo
 -------------------> sdf_path /hasan/Downloads/DEM/: LoadSDF, name = 47:48:351:352-hd, ret_val = -1
LoadSDF_SDF failed...assuming water level 0masl
Region  "47:48:351:352-hd" assumed as sea-level into page 1...returning from LoadSDF, loaded pages = 1
 topo loaded..printing dem attrs
alt =  -999
 [INFO] Remove shared memory segment { name = dem_attrs_shared_memory_segment }
 [INFO] Remove shared memory segment { name = dem_shared_memory_segment }

```

After pulling and recompiling, only 1 test now fails:

```shell
(openburst-venv2) hasan@hasans:~/openburst/openburst$ pytest --pyargs openburst
============================= test session starts =============================
rootdir: /home/hasan/openburst
configfile: pyproject.toml
collected 9 items

tests/db/test_db_conn.py .                                              [ 11%]
tests/dem/test_dem_folder.py .                                          [ 22%]
tests/dem/test_get_terrain_height.py F                                  [ 33%]
tests/openburst_splat/test_splat_load.py ...                            [ 66%]
tests/openburst_splat/test_splat_parallel_matrix.py .                   [ 77%]
tests/pcl/test_read_pcl_detections.py .                                 [ 88%]
tests/rad/test_rad.py .                                                 [100%]

================================== FAILURES ===================================
___________________________ test_get_terrain_height ___________________________

    def test_get_terrain_height():
        try:
            lat = 47.0
            lon = 8.1
            #lat = 57.41
            #lon = -6.19
            alt = geofunctions.get_terrain_height(splat,lat, lon)
>           assert alt > -160 # we make sure that there is some masl read. assuming -160masl the lowest location below sea level
E           assert -999 > -160

tests/dem/test_get_terrain_height.py:22: AssertionError

    def test_get_terrain_height():
        try:
            lat = 47.0
            lon = 8.1
            #lat = 57.41
            #lon = -6.19
            alt = geofunctions.get_terrain_height(splat,lat, lon)
            assert alt > -160 # we make sure that there is some masl read. assuming -160masl the lowest location below sea level
        except Exception: # pylint: disable=bare-except
>           pytest.fail("Could not read terrain height ..")
E           Failed: Could not read terrain height ..

tests/dem/test_get_terrain_height.py:24: Failed
---------------------------- Captured stdout call -----------------------------
:::::::: total DEM size = 3732480000
ok2: going to load topo
 -------------------> sdf_path /hasan/Downloads/DEM/: LoadSDF, name = 47:48:351:352-hd, ret_val = -1
LoadSDF_SDF failed...assuming water level 0masl
Region  "47:48:351:352-hd" assumed as sea-level into page 1...
=========================== short test summary info ===========================
FAILED tests/dem/test_get_terrain_height.py::test_get_terrain_height - Failed: Could not read terrain height ..
========================= 1 failed, 8 passed in 6.99s =========================
 [INFO] Remove shared memory segment { name = dem_attrs_shared_memory_segment }
 [INFO] Remove shared memory segment { name = dem_shared_memory_segment }

```

which is the terrain height. I modified the `test_get_terrain_height.py` to keep the original coordinates:
```shell
def test_get_terrain_height():
    try:
        lat = 47.0
        lon = 8.1
        #lat = 57.41
        #lon = -6.19
        alt = geofunctions.get_terrain_height(splat,lat, lon)
        assert alt > -160 # we make sure that there is some masl read. assuming -160mas>
    except Exception: # pylint: disable=bare-except
        pytest.fail("Could not read terrain height ..")
```

The files and contents seems to be okay:

```shell
(base) hasan@hasans:~/Downloads/DEM$ ls
44:45:348:349-hd.sdf  45:46:348:349-hd.sdf  46:47:348:349-hd.sdf  47:48:348:349-hd.sdf
44:45:349:350-hd.sdf  45:46:349:350-hd.sdf  46:47:349:350-hd.sdf  47:48:349:350-hd.sdf
44:45:350:351-hd.sdf  45:46:350:351-hd.sdf  46:47:350:351-hd.sdf  47:48:350:351-hd.sdf
44:45:351:352-hd.sdf  45:46:351:352-hd.sdf  46:47:351:352-hd.sdf  47:48:351:352-hd.sdf
44:45:352:353-hd.sdf  45:46:352:353-hd.sdf  46:47:352:353-hd.sdf  47:48:352:353-hd.sdf
44:45:353:354-hd.sdf  45:46:353:354-hd.sdf  46:47:353:354-hd.sdf  47:48:353:354-hd.sdf
(base) hasan@hasans:~/Downloads/DEM$ la -la 44:45:6:7:hd.sdf
ls: cannot access '44:45:6:7:hd.sdf': No such file or directory
(base) hasan@hasans:~/Downloads/DEM$ la -la 44:45:348:349-hd.sdf
-rw-rw-r-- 1 hasan hasan 27473921 Oct 23 12:40 44:45:348:349-hd.sdf
```

However, for each file conversion there was an EOF error thrown,
```shell
(base) hasan@hasans:~/Downloads$ chmod +x ./script.sh && ./script.sh
Converting /home/hasan/Downloads/L32/L32/N44E006.hgt
Reading /home/hasan/Downloads/L32/L32/N44E006.hgt...
*** Error: Premature EOF detected while reading "/home/hasan/Downloads/L32/L32/N44E006.hgt"!  :-(

Writing 44:45:353:354-hd.sdf... Done!
```
