### 2024-10-23 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir thank you. could you please paste here the line 247 of the header file "splatBurst.h" in folder "radterrain/SPLAT_RADIOPROP". That is the line that defines the DEM folder for splat and should look something like this:
`char burst_dem_path[] = "/home/user/Downloads/SDF_Files/";`
