### 2024-10-23 Swiss-Armed-Forces/openburst#1 @hasantahir
Boom,

Thanks @z-mathews. Indeed the path was set up incorrectly with the `/home/` prefix missing:

```shell
(openburst-venv2) hasan@hasans:~/openburst$ pytest --pyargs openburst
============================= test session starts =============================
rootdir: /home/hasan/openburst
configfile: pyproject.toml
collected 9 items

openburst/tests/db/test_db_conn.py .                                    [ 11%]
openburst/tests/dem/test_dem_folder.py .                                [ 22%]
openburst/tests/dem/test_get_terrain_height.py .                        [ 33%]
openburst/tests/openburst_splat/test_splat_load.py ...                  [ 66%]
openburst/tests/openburst_splat/test_splat_parallel_matrix.py .         [ 77%]
openburst/tests/pcl/test_read_pcl_detections.py .                       [ 88%]
openburst/tests/rad/test_rad.py .                                       [100%]

============================== 9 passed in 7.23s ==============================
 [INFO] Remove shared memory segment { name = dem_attrs_shared_memory_segment }
 [INFO] Remove shared memory segment { name = dem_shared_memory_segment }
```

I will now close this issue and move to exploring the core functionalities.
