### 2024-10-23 #7052 @hasantahir
I am happy with software after checking with the compile and trying out the functionalities.

@z-mathews can you please extend the paper to include how openBurst competes/performs against the current state of the field?
