### 2024-10-23 #7052 @rvg296
> I am happy with software after checking with the compile and trying out the functionalities.
>
> @z-mathews can you please extend the paper to include how openBurst competes/performs against the current state of the field?

@z-mathews  @hasantahir Yes, may be thats a great idea, especially comparing it against with current state of the art with some valid metrics in terms of performance computation and exploration.
