### 2024-10-24 Swiss-Armed-Forces/openburst#1 @z-mathews
@hasantahir thank you very much for running the installation, test and core functionality procedures of openBURST.
