### 2024-10-24 168ed6f
discussion added about related work in opensource air surveillance sensor simulation
