+++ docs/source/paper.bib
@@ -60,7 +60,7 @@
 
 @article{Mathews:2008,
   	author={Mathews, ~Z. and Bermudez i Badia, ~S. and Verschure, ~P.},
-  	jounral={International IEEE Conference Intelligent Systems}, 
+  	journal={International IEEE Conference Intelligent Systems}, 
 	title={Intelligent motor decision: From selective attention to a Bayesian world model}, 
 	year={2008},
 	volume={1},
@@ -69,3 +69,32 @@
 	doi={10.1109/IS.2008.4670418}
 }
 
+@article{Bluesky:2016,
+	author = {Hoekstra, ~J. ~M.  and Ellerbroek, ~J.}, 
+	title={BlueSky ATC Simulator Project: an Open Data and Open Source Approach}, 
+	journal = {Proceedings of the seventh International Conference for Research on Air Transport (ICRAT)}, 
+	year ={2016}
+}
+
+@article{Opensky:2014,
+	author = {Schäfer, M and Strohmeier, M and Lenders, V and Martinovic, I and and Wilhelm, M.},
+	title = {Bringing Up OpenSky: A Large-scale ADS-B Sensor Network for Research},
+	journal = {13th IEEE/ACM International Symposium on Information Processing in Sensor Networks (IPSN)}, 
+	pages = {83-94},
+	year = {2014}
+}
+
+@article{Stonesoup2017,
+	author = {Paul A. Thomas and Jordi Barr and Bhashyam Balaji and Kruger White},
+	title = {An open source framework for tracking and state estimation ('Stone Soup')},
+	volume = {10200},
+	booktitle = {Signal Processing, Sensor/Information Fusion, and Target Recognition XXVI},
+	editor = {Ivan Kadar},
+	organization = {International Society for Optics and Photonics},
+	publisher = {SPIE},
+	pages = {1020008},
+	keywords = {Tracking, State estimation, Open source, evaluation platform, algorithm repository},
+	year = {2017},
+	doi = {10.1117/12.2266249},
+	URL = {https://doi.org/10.1117/12.2266249}
+}
