+++ docs/source/paper.md
@@ -36,6 +36,11 @@ bibliography: paper.bib
 
 `openBURST` is a software project to design, deploy and test active/passive radar system networks for detecting simulated or recorded real world air traffic. It consists of decoupled software modules that can be replaced, extended or deployed independently for air surveillance sensor coverage and real-time detection computations. `openBURST` uses real-time communication between the distributed modules of the simulation framework, allowing for concurrent updates of target movements and sensor detections. Currently, `openBURST` supports coverage computation and real-time simulation of active radar and passive radar sensors (range-Doppler maps) using FM transmitters. The RF signal propagation, loss and terrain analysis tool [Splat!](https://www.qsl.net/kd2bd/splat.html), is extended for EM signal propagation computations with multi-core parallel processing and graphical user interfacing. `openBURST` uses [openstreetmap](https://openstreetmap.org) data with [openlayers](https://openlayers.org) for the interactive map. Terrain digital elevation data provided by [GMTED10](https://www.usgs.gov/coastal-changes-and-impacts/gmted2010) is used for Line-of-Sight and propagation loss computations. By implementing a client-server architecture, `openBURST` enables browser based clients to remain data and implementation agnostic. 
 
+Sensor simulation for air surveillance is of considerable interest both in the civilian and military fields as it allows for performance assessments of arbitrary sensor networks in diverse scenarios of interest. Such a simulation tool should be seen as the combination of three subparts: air traffic simulation, sensor detection simulation and sensor data fusion. 
+
+In the opensource domain, several tools exist for the subpart of air traffic simulation. E.g. BlueSky [@Bluesky:2016] presents a fully open-source and open-data approach to air traffic simulation, whereas the
+OpenSky network [@Opensky:2014] is a non-profit community-based receiver network that collects air traffic surveillance data and makes it accessible to academic and institutional researchers. In the domain of air surveillance sensor detection simulation, opensource tools for passive radar simulation such as [PassiveRadarSim](https://github.com/bradleeharr/PassiveRadarSim) and for active radar simulation such as [radarsimpy](https://github.com/radarsimx/radarsimpy) have been published. In the active radar domain, sophisticated opensource tools for very specific use cases such as for Synthetic Aperture Radar (SAR), e.g. [RaySAR](https://github.com/StefanJAuer/RaySAR) or variats of generic wave propagation tools such as [gprMAX](https://github.com/gprMax/gprMax) also exist. Diverse data fusion, tracking and state estimation tools have also been opensourced such as Stonesoup [@Stonesoup2017]. However, currently no extendable real-time simulation suite for air traffic simulation and sensor detection for active/passive sensors that incorporates wave propagation modelling is available to the opensource community. `openBURST`is an attempt to fill this gap, proving the research and air surveillance simulation community with a general-purpose and modular tool, that provides a user-friendly browser based interface. 
+
[…]
