### 2024-10-24 53df424
new discussion typos corrected
