### 2024-10-24 #7052 @z-mathews
@hasantahir @rvg296 thank you for the suggestion on updating the paper. The section "statement of need" in the paper has been extended with references to currently available opensource tools for air traffic simulation, sensor detection simulation and data fusion and tracking/state estimation. The added text discussed which gap `openBURST` intends to fill.
