### 2024-10-24 #7052 @danielskatz
👋 @rvg296

 > I am happy with software after checking with the compile and trying out the functionalities.

Can you update [your checklist](https://github.com/openjournals/joss-reviews/issues/7052#issuecomment-2273735993) as appropriate?
