### 2024-10-28 d2adf95
removed np.float64 type error in replay data
