### 2024-10-28 0d2cc81
removed replay ref target np.float64 error
