### 2024-11-12 #7052 @danielskatz
👋 @rvg296 - We're now down to the minimum of two reviewers, so I hope we can count on you to complete your review.  I think you've done more than I see.

Can you update [your checklist](https://github.com/openjournals/joss-reviews/issues/7052#issuecomment-2273735993) as appropriate? And let us know what else is blocking you from further progress?
