### 2024-11-24 #7052 @danielskatz
@rvg296 - Can you provide an update on your last couple of checklist items?
