### 2024-11-27 b71bd03
small changes in paper.bib
