+++ docs/source/paper.bib
@@ -10,13 +10,13 @@
   	url = {https://www.frontiersin.org/articles/10.3389/frsip.2022.822894/full},
   	Author = {{Mathews}, ~Z. and {Quiriconi}, ~L. and {Schuepbach}, ~C. and {Weber}, ~P.},
     Journal = {Frontiers in Signal Processing, Volume 2},
-  	Title = {{Learning Resource Allocation in Active-Passive Radar Sensor Networks}},
+  	Title = {Learning Resource Allocation in Active-Passive Radar Sensor Networks},
   	Year = 2022
 }
 
 @article{Mousel:2017,
     Author = {{Mousel}, ~P},
-    title = "{Passive Radar Coverage Optimization}",
+    title = {Passive Radar Coverage Optimization},
     journal = {Master Thesis, Institute of Electromagnetic Fields (IEF), ETH Zurich},
     Year = 2017
 }
@@ -24,7 +24,7 @@
 @article{Studer:2022,
 	url = {https://ieeexplore.ieee.org/abstract/document/9764225},
 	Author = {{Studer}, ~S. and {Mathews}, ~Z. and {Welschen}, ~S. and {Schuepbach} ~C. and {Leuthold}, ~J. and {Smajic}, ~S.},
-	Title = {{Maximum Information Fusion of Passive and Active Radars for Air Surveillance}}, 
+	Title = {Maximum Information Fusion of Passive and Active Radars for Air Surveillance}, 
 	journal = {IEEE Radar Conference (RadarConf), New York City, NY, USA,pp. 1-6},
 	Year = 2022
 }
@@ -61,7 +61,7 @@
 @article{Mathews:2008,
   	author={Mathews, ~Z. and Bermudez i Badia, ~S. and Verschure, ~P.},
   	journal={International IEEE Conference Intelligent Systems}, 
-	title={Intelligent motor decision: From selective attention to a Bayesian world model}, 
+	title={Intelligent motor decision: From selective attention to a {B}ayesian world model}, 
 	year={2008},
 	volume={1},
 	pages={4-8-4-13},
@@ -86,7 +86,7 @@
 
 @article{Stonesoup2017,
 	author = {Paul A. Thomas and Jordi Barr and Bhashyam Balaji and Kruger White},
-	title = {An open source framework for tracking and state estimation ('Stone Soup')},
+	title = {An open source framework for tracking and state estimation ('{S}tone {S}oup')},
 	volume = {10200},
 	booktitle = {Signal Processing, Sensor/Information Fusion, and Target Recognition XXVI},
 	editor = {Ivan Kadar},
