### 2024-11-27 #7052 @danielskatz
👋 @z-mathews - I've also found a number of small issues, as indicated in https://github.com/Swiss-Armed-Forces/openburst/pull/4 - you might want to merge this before working on the DOIs ☝️ to avoid a possible merge conflict.
