### 2024-11-28 02586a8
removed pcl detection sqllite logging errors
