### 2024-11-28 bb1f72a
missing DOIs added; all DOI suggestions by joss editorialbot were correct.
