+++ docs/source/paper.bib
@@ -2,7 +2,8 @@
   	url = {https://ieeexplore.ieee.org/document/7131222/},
   	Author = {{Mathews}, ~Z. and {Quiriconi}, ~L. and {Weber}, ~P.},
   	Journal = {2015 IEEE Radar Conference (RadarCon)},
-  	Title = {{Multi-static passive receiver location optimization in alpine terrain using a parallelized genetic algorithm}},
+  	Title = {Multi-static passive receiver location optimization in alpine terrain using a parallelized genetic algorithm},
+	doi = {10.1109/RADAR.2015.7131222},
   	Year = 2015
 }
 
@@ -11,6 +12,7 @@
   	Author = {{Mathews}, ~Z. and {Quiriconi}, ~L. and {Schuepbach}, ~C. and {Weber}, ~P.},
     Journal = {Frontiers in Signal Processing, Volume 2},
   	Title = {Learning Resource Allocation in Active-Passive Radar Sensor Networks},
+	doi = {10.3389/frsip.2022.822894},
   	Year = 2022
 }
 
@@ -26,6 +28,7 @@
 	Author = {{Studer}, ~S. and {Mathews}, ~Z. and {Welschen}, ~S. and {Schuepbach} ~C. and {Leuthold}, ~J. and {Smajic}, ~S.},
 	Title = {Maximum Information Fusion of Passive and Active Radars for Air Surveillance}, 
 	journal = {IEEE Radar Conference (RadarConf), New York City, NY, USA,pp. 1-6},
+	doi = {10.1109/RadarConf2248738.2022.9764225},
 	Year = 2022
 }
 
@@ -81,6 +84,7 @@
 	title = {Bringing Up OpenSky: A Large-scale ADS-B Sensor Network for Research},
 	journal = {13th IEEE/ACM International Symposium on Information Processing in Sensor Networks (IPSN)}, 
 	pages = {83-94},
+	doi = {10.1109/IPSN.2014.6846743},
 	year = {2014}
 }
 
