### 2024-11-28 #7052 @z-mathews
@danielskatz thank you for pointing out the small issues in the paper; the corresponding pull request was merged. Also the DOIs suggested by the @editorialbot were correct and the .bib file items were extended accordingly and pushed to the joss branch.
