### 2024-11-28 708fb6c
extenden pcl detection logging with target heading and speed in the ground truth
