### 2024-11-28 #7052 @danielskatz
Congratulations to @z-mathews (Zenon Mathews) and co-authors on your publication!!

And thanks to @rvg296, @hasantahir for reviewing!
JOSS depends on volunteers and we couldn't succeed without you
