### 2024-11-30 #7052 @z-mathews
@danielskatz @hasantahir @rvg296 Thank you very much for the review process. openBURST documentation and installation process etc. improved remarkably thanks to the JOSS review process! I hope you don't mind that we mention you in the Acknowledgements sections of [openBURST GitHub repo ](https://github.com/Swiss-Armed-Forces/openburst)
and the [read-the-docs documentation](https://openburst.readthedocs.io/en/latest/README_RTD.html#acknowledgments).
